refactor: Move provider-specific code into provider plugins
- Move Codex-only code into codex.ts (module-private):
  buildCodexInterruptRequest, buildCodexApprovalResponse,
  CODEX_MODEL_LABELS, CODEX_PERMISSION_MODE_LABELS,
  DEFAULT_CODEX_MODEL, DEFAULT_CODEX_EFFORT

- Move Claude Code-only code into claude.tsx (module-private):
  buildInterruptRequest, generateRandomId,
  ClaudeCodeSettingsPanel, ClaudeCodeTriggerLabel

- Move codexRenderers.tsx into providers/ directory

- Add defaultModel/defaultEffort to ProviderPlugin interface;
  replace if-else in defaultModelForProvider/defaultEffortForProvider
  with plugin dispatch

- Extract settingsShared.tsx for RadioGroup and label helpers
  shared between EditorSettingsDropdown and claude.tsx plugin

- controlResponse.ts reduced to generic utilities only — no
  provider-specific constants, builders, or labels remain
